added profilePageUtils.js and inserted functions imported functions to user.js
