+++ src/user/index.html
@@ -5,7 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Social</title>
     <script type="module" src="../lib/main.js"></script>
-    <script type="module" src="../lib/page_scripts/profile.js"></script>
+    <script type="module" src="../lib/page_scripts/user.js"></script>
     <script
       src="https://kit.fontawesome.com/a6b44b4960.js"
       crossorigin="anonymous"
